+++ executor/executor.go
@@ -30,6 +30,10 @@ func Run(cfg ExecutorConfig) {
 		numWorkers = len(cnf.Scripts)
 	}
 
+	if len(cnf.SeriesPrefix) >= 1 {
+		lib.SetSeriesPrefix(cnf.SeriesPrefix)
+	}
+
 	work := NewWorkQueue(4, numWorkers, len(cnf.Scripts))
 	log.Info("Starting script execution workers...")
 	work.Process()
@@ -62,6 +66,16 @@ func Run(cfg ExecutorConfig) {
 				Value: 1.0,
 				Type:  "gauge",
 				Help:  "indicates that a script has been identified to be executed",
+			})
+
+			scriptLoadedSeries = append(scriptLoadedSeries, lib.Metric{
+				Name: "script_total_execution_time_ms",
+				Labels: map[string]string{
+					"script": res.ScriptPath,
+				},
+				Value: float64(res.TotalExecTime),
+				Type:  "gauge",
+				Help:  "indicates the number of milliseconds it has taken to execute the script",
 			})
 
 			lastRunSuccess := 0.0
@@ -105,7 +119,11 @@ func Run(cfg ExecutorConfig) {
 
 	seriesOutput := lib.GenerateSeries(series, execSuccess)
 
-	log.Infof("Writing resulting series to %s", cfg.OutputFilePath)
+	if cfg.OutputFilePath != "" {
+		log.Infof("Writing resulting series to %s", cfg.OutputFilePath)
+	} else {
+		log.Info("Writing resulting series to stdout")
+	}
 	if !cfg.Simulate {
 		if len(series) >= 1 {
 			// Write the series to the output file