+++ lib/lib.go
@@ -6,11 +6,13 @@ import (
 	"math"
 	"path/filepath"
 	"regexp"
+	"runtime"
 	"strconv"
 	"strings"
 	"time"
 
 	"github.com/google/renameio"
+	"github.com/hartfordfive/n2p-script-executor/version"
 
 	//"github.com/hartfordfive/n2p-script-executor/executor"
 	log "github.com/sirupsen/logrus"
@@ -25,11 +27,20 @@ type Metric struct {
 	Help   string
 }
 
+var seriesPrefix = "n2p_script_exec"
+
+// SetSeriesPrefix updates the prefix of the created series.
+func SetSeriesPrefix(prefix string) {
+	if len(prefix) >= 1 {
+		seriesPrefix = prefix
+	}
+}
+
 func (m Metric) String(addHelp bool) string {
 	output := ""
 	if addHelp {
-		output += fmt.Sprintf("# TYPE n2p_script_exec_%s %s\n", m.Name, m.Type)
-		output += fmt.Sprintf("# HELP n2p_script_exec_%s %s\n", m.Name, m.Help)
+		output += fmt.Sprintf("# TYPE %s_%s %s\n", seriesPrefix, m.Name, m.Type)
+		output += fmt.Sprintf("# HELP %s_%s %s\n", seriesPrefix, m.Name, m.Help)
 	}
 	if len(m.Labels) >= 1 {
 		flattendLabels := []string{}
@@ -37,21 +48,39 @@ func (m Metric) String(addHelp bool) string {
 			flattendLabels = append(flattendLabels, fmt.Sprintf("%s=\"%s\"", k, v))
 		}
 		if ValueCanBeInt(m.Value) {
-			output += fmt.Sprintf("n2p_script_exec_%s{%s} %s", m.Name, strings.Join(flattendLabels, ", "), convertToIntString(m.Value))
+			output += fmt.Sprintf("%s_%s{%s} %s", seriesPrefix, m.Name, strings.Join(flattendLabels, ", "), convertToIntString(m.Value))
 		} else {
-			output += fmt.Sprintf("n2p_script_exec_%s{%s} %f", m.Name, strings.Join(flattendLabels, ", "), m.Value)
+			output += fmt.Sprintf("%s_%s{%s} %f", seriesPrefix, m.Name, strings.Join(flattendLabels, ", "), m.Value)
 		}
 
 	} else {
 
 		if ValueCanBeInt(m.Value) {
-			output += fmt.Sprintf("n2p_script_exec_%s %s", m.Name, convertToIntString(m.Value))
+			output += fmt.Sprintf("%s_%s %s", seriesPrefix, m.Name, convertToIntString(m.Value))
 		} else {
-			output += fmt.Sprintf("n2p_script_exec_%s %f", m.Name, m.Value)
+			output += fmt.Sprintf("%s_%s %f", seriesPrefix, m.Name, m.Value)
 		}
 	}
 	//output += "\n"
 	return output
+}
+
+// IsValidMetricName returns if a metric name is valid or not
+func (m Metric) IsValidMetricName() bool {
+	if match, _ := regexp.MatchString("[a-zA-Z_:][a-zA-Z0-9_:]*", m.Name); match {
+		return true
+	}
+	return false
+}
+
+// ValidSeriesLabels returns if a metric name is valid or not
+func (m Metric) ValidSeriesLabels() bool {
+	for lblName := range m.Labels {
+		if match, _ := regexp.MatchString("[a-zA-Z_][a-zA-Z0-9_]*", lblName); !match {
+			return false
+		}
+	}
+	return true
 }
 
 // GetScriptName returns the script name without the extension
@@ -136,23 +165,43 @@ func GenerateSeries(metrics []Metric, execSuccess []string) string {
 			typeHelpLine[metric.Name] = true
 		}
 		log.Debugf("Adding metric: %s", metric.String(addHelp))
-		data += metric.String(addHelp) + "\n"
-		addHelp = false
+		if !metric.IsValidMetricName() {
+			log.Warnf("Metric %s has an invalid name. Skipping it.", metric.Name)
+		} else if !metric.ValidSeriesLabels() {
+			log.Warnf("Metric %s has invalid labels. Removing labels.")
+			metric.Labels = map[string]string{}
+		} else {
+			log.Debugf("Metric name '%s' is valid!", metric.Name)
+			data += metric.String(addHelp) + "\n"
+			addHelp = false
+		}
 	}
 
-	data += "# TYPE n2p_script_exec_lastrun counter\n"
-	data += "# HELP n2p_script_exec_lastrun Time when the script was last executed\n"
+	data += fmt.Sprintf("# TYPE %s_lastrun counter\n", seriesPrefix)
+	data += fmt.Sprintf("# HELP %s_lastrun Time when the script was last executed\n", seriesPrefix)
 	for _, script := range execSuccess {
 		data += generateScriptCheckpointMetric(script) + "\n"
 	}
 	data += GenerateExecutorCheckpointMetric() + "\n"
+
+	data += fmt.Sprintf("# TYPE %s_build_info counter\n", seriesPrefix)
+	data += fmt.Sprintf("# HELP %s_build_info Build information of the script executor\n", seriesPrefix)
+	data += fmt.Sprintf(
+		"%s_build_info{version=\"%s\", commit_hash=\"%s\", build_date=\"%s\", go_version=\"%s\"} 1",
+		seriesPrefix,
+		version.Version,
+		version.CommitHash,
+		version.BuildDate,
+		runtime.Version(),
+	)
 
 	return data
 }
 
 func generateScriptCheckpointMetric(scriptName string) string {
 	return fmt.Sprintf(
-		"n2p_script_exec_lastrun{script=\"%s\"} %d",
+		"%s_lastrun{script=\"%s\"} %d",
+		seriesPrefix,
 		scriptName,
 		(time.Now().UnixNano() / int64(time.Millisecond)),
 	)
@@ -161,7 +210,8 @@ func generateScriptCheckpointMetric(scriptName string) string {
 // GenerateExecutorCheckpointMetric is ...
 func GenerateExecutorCheckpointMetric() string {
 	return fmt.Sprintf(
-		"n2p_script_exec_last_execution %d",
+		"%s_last_execution %d",
+		seriesPrefix,
 		(time.Now().UnixNano() / int64(time.Millisecond)),
 	)
 }